SAML-19 Allow configurable AuthnContext for IdPs
When crafting an authn request for an IdP, we as a service provider
can specify an "AuthnContext" that specifies the authentication
mechanism we expect the IdP to use when authenticating. For example,
this could be basic username and password auth. We can provide
both a URN of an authentication mechanism AND a comparison operator.

Until now, we have been hardcoding an "EXACT" comparison operator
and the "PasswordProtectedTransport" class. This change makes the
new default operator of "MINIMUM". We also now allow developers
to specify these values in their configuration of an IdP by adding
acClassComparison and acClassRef config keys.